Fix winport.pl on newer Perl versions, update README
Using defined in this manner has been deprecated for many versions, but
now returns an error in the newest versions of Perl. The README makes no
mention of needing XML::LibXML in run the script properly, and despite
it being mentioned in the original UnlinkMKV repository, is warranted in
being reiterated here.